% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnv.R
\name{get_annotations}
\alias{get_annotations}
\title{Get gene annotations}
\usage{
get_annotations(org = "human", version = "latest")
}
\arguments{
\item{org}{The organism to get annotations for. Can be "human" or "mouse".}

\item{version}{The version of the annotations to use. Can be "latest" or a specific version number.}
}
\value{
A data frame containing the gene annotations.
}
\description{
This function retrieves the gene annotations for a given organism and version.
}
\examples{
\dontrun{
get_annotations(org = "human", version = "latest")
}
}
