% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnv.R
\name{run_infercnv_individual}
\alias{run_infercnv_individual}
\title{Run infercnv on an individual object}
\usage{
run_infercnv_individual(
  obj,
  ref_obj = NULL,
  individual_name,
  annot_column,
  org,
  version = "latest",
  assay = "RNA",
  save_dir = getwd(),
  ncores = 1,
  ...
)
}
\arguments{
\item{obj}{The object to run infercnv on.}

\item{ref_obj}{The reference object to run infercnv on.}

\item{individual_name}{The name of the individual to run infercnv on.}

\item{annot_column}{The column name of the annotations to run infercnv on.}

\item{org}{The organism to run infercnv on.}

\item{version}{The version of the annotations to use.}

\item{assay}{The assay to run infercnv on.}

\item{save_dir}{The directory to save the infercnv results.}

\item{ncores}{The number of cores to use.}

\item{...}{Additional arguments to pass to infercnv::run.}
}
\value{
A list of infercnv objects.
}
\description{
This function runs infercnv on an individual object.
}
\examples{
\dontrun{
run_infercnv_individual(obj, ref_obj = NULL, individual_name, annot_column, org = "human", version = "latest", assay = "RNA", save_dir = getwd(), ncores = 1, ...)
}
}
