% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infercnv.R
\name{run_infercnv_multi}
\alias{run_infercnv_multi}
\title{Run infercnv on multiple objects}
\usage{
run_infercnv_multi(
  obj,
  split.by,
  ref_obj = NULL,
  org,
  annot_column,
  save_dir = getwd(),
  ...
)
}
\arguments{
\item{obj}{The object to run infercnv on.}

\item{split.by}{The column name to split the object by.}

\item{ref_obj}{The reference object to run infercnv on.}

\item{org}{The organism to run infercnv on.}

\item{save_dir}{The directory to save the infercnv results.}

\item{...}{Additional arguments to pass to run_infercnv_individual.}
}
\description{
This function runs infercnv on multiple objects.
}
\examples{
\dontrun{
run_infercnv_multi(obj, split.by, ref_obj = NULL, org = "human", annot_column, save_dir = getwd(), ...)
}
}
